Split Point2D and Point3D into different files
